+++ libraries/core/src/index.ts
@@ -3,3 +3,4 @@ export * from './types';
 export type { Database } from './database.types';
 export * from './sdk';
 export * from './date';
+export * from './classifyLocation';